+++ src/extension.ts
@@ -40,6 +40,9 @@ import { getSlackItems } from "./utils/treeview/getSlackItems";
 import { getCodeContextSummary } from "./utils/treeview/getCodeContextSummary";
 import setLoading from "./utils/vscode/setLoading";
 import addToRecommendedCommandHandler from "./utils/commands/addToRecommended";
+import ghHover from "./utils/components/ghHover";
+import slackHover from "./utils/components/slackHover";
+import jiraHover from "./utils/components/jiraHover";
 
 // repo information
 let owner: string | undefined = "";
@@ -224,16 +227,6 @@ export class WatermelonTreeDataProvider
 
 export async function activate(context: vscode.ExtensionContext) {
   setLoggedIn(false);
-  if (vscode.env.uiKind === vscode.UIKind.Web) {
-    //log all env variables for web
-    console.log("vscode.env.appName", vscode.env.appName);
-    console.log("vscode.env.appRoot", vscode.env.appRoot);
-    console.log("vscode.env.appUriScheme", vscode.env.appHost);
-    console.log("vscode.env.remoteName", vscode.env.remoteName);
-    console.log(vscode.Uri.parse(vscode.env.uriScheme));
-    console.log("vscode.env.appUriScheme", vscode.env.uriScheme);
-    console.log(process.env);
-  }
 
   // allows saving state across sessions
   const workspaceState: { repo: string; owner: string } | undefined =
@@ -314,7 +307,10 @@ export async function activate(context: vscode.ExtensionContext) {
   updateStatusBarItem(wmStatusBarItem);
 
   // create the hover provider
-  let wmHover = hover({ reporter });
+  hover({ reporter });
+  ghHover({ reporter });
+  slackHover({ reporter });
+  jiraHover({ reporter });
 
   let loginCommandHandler = async () => {
     // Get our PAT session.